backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
pydantic[email]
APScheduler
pytz
yagmail
keyring
